Move Viewport animations to Mapsui.Animations folder
