+++ Mapsui/Animations/ZoomAroundLocationAnimation.cs
@@ -1,8 +1,7 @@
 ﻿using System.Collections.Generic;
-using Mapsui.Animations;
 using Mapsui.Utilities;
 
-namespace Mapsui.ViewportAnimations;
+namespace Mapsui.Animations;
 
 internal class ZoomAroundLocationAnimation
 {
