+++ Mapsui/Viewport.cs
@@ -11,7 +11,6 @@ using System.Runtime.CompilerServices;
 using Mapsui.Animations;
 using Mapsui.Extensions;
 using Mapsui.Limiting;
-using Mapsui.ViewportAnimations;
 
 namespace Mapsui;
